# server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gem Auto Rentals API
  description: |
    RESTful API for Gem Auto Rentals - a car rental management platform.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Roles
    - **CUSTOMER**: Regular users who can book vehicles
    - **SUPPORT**: Staff who can view and manage bookings
    - **MANAGER**: Staff with additional fleet management access
    - **ADMIN**: Full system access

    ## Rate Limiting
    API requests are rate limited to 100 requests per 15 minutes per IP address.
  version: 1.0.0
  contact:
    name: Gem Auto Rentals Support
    email: support@gemautorentals.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.gemautorentals.com/api
    description: Production server
...
paths:
...
  /payments/create-intent:
    post:
      tags: [Payments]
      summary: Create payment intent
      description: Create a Stripe payment intent for a booking
      operationId: createPaymentIntent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId]
              properties:
                bookingId:
                  type: string
                  description: CUID of the booking to pay for
            example:
              bookingId: clm1234567890
      responses:
        '200':
          description: Payment intent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      clientSecret:
                        type: string
                        description: Stripe client secret for frontend
                      amount:
                        type: number
                        description: Payment amount in dollars
        '400':
          description: Payment processing not configured or booking already paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/confirm:
    post:
      tags: [Payments]
      summary: Confirm payment
      description: Confirm payment after Stripe processing
      operationId: confirmPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paymentIntentId]
              properties:
                paymentIntentId:
                  type: string
            example:
              paymentIntentId: pi_1234567890
      responses:
        '200':
          description: Payment confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      payment:
                        $ref: '#/components/schemas/Payment'
                      bookingStatus:
                        $ref: '#/components/schemas/BookingStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/{bookingId}:
    get:
      tags: [Payments]
      summary: Get payment details
      description: Get payment information for a booking
      operationId: getPayment
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/{bookingId}/refund:
    post:
      tags: [Payments]
      summary: Process refund
      description: Process a full or partial refund (staff only)
      operationId: refundPayment
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  description: Refund amount (defaults to full refund)
                reason:
                  type: string
                  description: Reason for refund
      responses:
        '200':
          description: Refund processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      payment:
                        $ref: '#/components/schemas/Payment'
                      refund:
                        type: object
                        properties:
                          id:
                            type: string
                          amount:
                            type: number
                          status:
                            type: string
        '400':
          description: Cannot refund - payment not successful or no Stripe payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Unauthorized
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Not found
  schemas:
...
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, ACTIVE, COMPLETED, CANCELLED]
    PaymentStatus:
      type: string
      enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, REFUNDED, PARTIALLY_REFUNDED]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
    Payment:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        amount:
          type: number
        status:
          $ref: '#/components/schemas/PaymentStatus'
        method:
          type: string
          enum: [CARD, CASH, BANK_TRANSFER]
          nullable: true
        stripePaymentIntentId:
          type: string
          nullable: true
        stripeChargeId:
          type: string
          nullable: true
        refundAmount:
          type: number
          nullable: true
        refundReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time